+++ environments/gloo-platform/gwapi-mgmt-gm-workers/cluster2/gloo-platform/init.sh
@@ -7,7 +7,7 @@ echo "deploy and register gloo-mesh agent and addons"
 if [[ ${gloo_mesh_version} == "" ]]
   then
     # provide gloo_mesh_version variable
-    echo "Please provide the gloo_mesh_version to use (i.e. 2.7.0-beta0-2024-10-16-puertomontt-gg-env-500a30fe59):"
+    echo "Please provide the gloo_mesh_version to use (i.e. 2.7.0-beta0-2024-10-24-main-389ae0d3be):"
     read gloo_mesh_version
 fi
 
@@ -75,7 +75,7 @@ spec:
                   namespace: null
                   
     repoURL: https://storage.googleapis.com/gloo-platform-dev/platform-charts/helm-charts
-    targetRevision: 2.7.0-beta0-2024-10-16-puertomontt-gg-env-500a30fe59
+    targetRevision: 2.7.0-beta0-2024-10-24-main-389ae0d3be
   syncPolicy:
     automated:
       prune: true
@@ -123,7 +123,7 @@ spec:
             image:
               pullPolicy: IfNotPresent
               repository: gcr.io/gloo-mesh/gloo-otel-collector
-              tag: 2.7.0-beta0-2024-10-16-puertomontt-gg-env-500a30fe59
+              tag: 2.7.0-beta0-2024-10-24-main-389ae0d3be
             config:
                 exporters:
                     otlp:
@@ -145,7 +145,7 @@ spec:
                   name: cilium-run
                   
     repoURL: https://storage.googleapis.com/gloo-platform-dev/platform-charts/helm-charts
-    targetRevision: 2.7.0-beta0-2024-10-16-puertomontt-gg-env-500a30fe59
+    targetRevision: 2.7.0-beta0-2024-10-24-main-389ae0d3be
   syncPolicy:
     automated:
       prune: true